stats/pyproject.toml
[tool.poetry]
name = "stats"
version = "0.1.0"
description = ""
authors = ["Andrey Sobolev <email.asobolev@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "0.28.0"
fastapi = { extras = ["all"], version = "0.100.0" }
uvicorn = { extras = ["standard"], version = "0.23.1" }
sqlalchemy = "2.0.19"
alembic = "1.11.1"
orjson = "3.9.2"
loguru = "0.7.0"
pydantic-settings = "2.0.2"
aiokafka = "^0.8.1"
simple-print = "^1.7.5"
rich = "^13.4.1"

[tool.poetry.dev-dependencies]
pytest = "7.4.0"
mypy = "1.4.1"
pre-commit = "3.3.3"
pytest-asyncio = "0.21.1"
pytest-alembic = "0.10.6"
pytest-cov = "4.1.0"
ruff = "0.0.278"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v"
pythonpath = [
  ".",
]
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.ruff]
select = ["E", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "migrations"
]
per-file-ignores = {}
line-length = 88
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"
